moved all WithTheme tests to single file
